+++ example/async_viewmodel_flutter_di/lib/di/service_config.dart
@@ -1,0 +1,12 @@
+import 'package:force_inject/force_inject.dart';
+import '../viewmodels/auth_view_model.dart';
+
+final services = ServiceCollection()
+  ..addScoped<AuthViewModel, AuthViewModel>();
+
+void configureConstructors() {
+  ServiceProvider.registerConstructor<AuthViewModel>(
+        () => AuthViewModel(),
+    [],
+  );
+}