+++ packages/force_inject_flutter/lib/force_inject_flutter.dart
@@ -4,3 +4,4 @@ library force_inject_flutter;
 export 'src/force_scope_widget.dart';
 export 'src/force_view_model_builder.dart';
 export 'src/force_initializable_view_model.dart';
+export 'src/force_async_initializable_view_model.dart';