+++ packages/force_inject_flutter/lib/src/force_async_initializable_view_model.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/widgets.dart';
+
+abstract class ForceAsyncInitializableViewModel {
+  Future<void> initAsync(BuildContext context);
+}